+++ src/lib/HST/Frontend/Transformer/Messages.hs
@@ -26,7 +26,17 @@ skipNotSupported
   => String      -- ^ The name of the feature (plural) that is not supported.
   -> S.SrcSpan a -- ^ The source span of the excerpt to display.
   -> Sem r ()
-skipNotSupported feature srcSpan = report
-  $ message Info srcSpan
-  $ feature ++ " are not supported and will be skipped!"
+skipNotSupported feature srcSpan
+  = skipNotSupported' feature "will be skipped" srcSpan
 
+-- | Like 'skipNotSupported' but has an additional parameter that specifies how
+--   the AST node will be skipped.
+skipNotSupported'
+  :: Member Report r
+  => String      -- ^ The name of the feature (plural) that is not supported.
+  -> String      -- ^ The strategy that is applied to skip the node.
+  -> S.SrcSpan a -- ^ The source span of the excerpt to display.
+  -> Sem r ()
+skipNotSupported' feature strategy srcSpan = report
+  $ message Info srcSpan
+  $ feature ++ " are not supported and " ++ strategy ++ "!"